Update a tweet dataset with last information

Update information of all statuses listed. At most, the last 90000.

update_data(
  x,
  path,
  statuses,
  token = NULL,
  return_all = FALSE,
  overwrite = TRUE
)

Arguments

x

dataset to update. Not used if path is set.

path

Path to .rds file with saved dataset

statuses

Tweet statuses to update

token

Expert use only. Use this to override authentication for a single API call. In most cases you are better off changing the default for all calls. See auth_as() for details.

return_all

Logical. TRUE returns the complete dataset, FALSE returns data in statuses list (default).

overwrite

Logical. Whether to overwrite rds file with updated data